fixed PosixMutex impl, moved gl context creation to separate function and removed io.h and fcntl.h from Libs
