+++ Includes/Utility/Libs.h
@@ -4,8 +4,6 @@
    ======================================================================== */
 #include <string.h>
 #include <stdio.h>
-#include <io.h>
-#include <fcntl.h>
 #include <vector>
 #include <deque>
 #include <stdarg.h>
